requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg[binary]
python-jose
passlib[bcrypt]
pydantic
email-validator
httpx
python-dotenv
